workspace/library/templates/ai-stripe-agent/pyproject.toml
[project]
name = "ai-stripe-agent"
version = "1.0.0"
description = "AI Agent for Stripe Payment Operations"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "ULTRA-CREATE", email = "ultra-create@example.com" }
]
keywords = ["ai", "agent", "stripe", "payments", "phidata"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
]

dependencies = [
    "phidata>=2.0.0",
    "stripe>=7.0.0",
    "stripe-agent-toolkit>=0.1.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
]
anthropic = [
    "anthropic>=0.25.0",
]

[project.scripts]
stripe-agent = "src.agent:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
